Update stale source tree

+++ bcm/devarch.c
@@ -76,8 +76,8 @@ archwalk(Chan* c, Chan *nc, char** name, int nname)
 	return devwalk(c, nc, name, nname, archdir, narchdir, devgen);
 }
 
-static long
-archstat(Chan* c, uchar* dp, long n)
+static int
+archstat(Chan* c, uchar* dp, int n)
 {
 	return devstat(c, dp, n, archdir, narchdir, devgen);
 }
